refactor: extract non-streaming proxy handler
