+++ src/routes/shared/non-streaming-handler.ts
@@ -1,0 +1,244 @@
+import type { Context } from "hono";
+import { CodexApiError } from "../../proxy/codex-api.js";
+import type { CodexApi, WsPoolContext } from "../../proxy/codex-api.js";
+import type { AccountPool } from "../../auth/account-pool.js";
+import type { CookieJar } from "../../proxy/cookie-jar.js";
+import type { ProxyPool } from "../../proxy/proxy-pool.js";
+import { EmptyResponseError, UpstreamPrematureCloseError } from "../../translation/codex-event-extractor.js";
+import { withRetry } from "../../utils/retry.js";
+import { acquireAccount, releaseAccount } from "./account-acquisition.js";
+import { toErrorStatus } from "./proxy-error-handler.js";
+import { enqueueLogEntry } from "../../logs/entry.js";
+import { recordStreamCloseEvent } from "../../logs/stream-close-event.js";
+import type { SessionAffinityMap } from "../../auth/session-affinity.js";
+import type { FormatAdapter, ProxyRequest, UsageHint } from "./proxy-handler.js";
+import { annotateImageGenOutcome, buildCodexApi, stripCodexErrorPrefix } from "./proxy-handler-utils.js";
+
+const MAX_EMPTY_RETRIES = 2;
+
+export interface HandleNonStreamingOptions {
+  c: Context;
+  accountPool: AccountPool;
+  cookieJar?: CookieJar;
+  req: ProxyRequest;
+  fmt: FormatAdapter;
+  proxyPool?: ProxyPool;
+  initialApi: CodexApi;
+  initialResponse: Response;
+  initialEntryId: string;
+  abortController: AbortController;
+  released: Set<string>;
+  requestId: string;
+  affinityMap?: SessionAffinityMap;
+  conversationId?: string | null;
+  turnState?: string;
+  getUsageHint?: () => UsageHint | undefined;
+  restoreImplicitResumeRequest?: () => void;
+  buildPoolCtx?: (forEntryId: string) => WsPoolContext | undefined;
+  setActiveAccount?: (entryId: string, api: CodexApi) => void;
+  variantHash?: string;
+}
+
+export async function handleNonStreaming(options: HandleNonStreamingOptions): Promise<Response> {
+  const {
+    c,
+    accountPool,
+    cookieJar,
+    req,
+    fmt,
+    proxyPool,
+    initialApi,
+    initialResponse,
+    initialEntryId,
+    abortController,
+    released,
+    requestId,
+    affinityMap,
+    conversationId,
+    turnState,
+    getUsageHint,
+    restoreImplicitResumeRequest,
+    buildPoolCtx,
+    setActiveAccount,
+    variantHash,
+  } = options;
+  let currentEntryId = initialEntryId;
+  let currentApi = initialApi;
+  let currentRawResponse = initialResponse;
+
+  for (let attempt = 1; ; attempt++) {
+    try {
+      const responseFunctionCallIds = new Set<string>();
+      const result = await fmt.collectTranslator({
+        api: currentApi,
+        response: currentRawResponse,
+        model: req.model,
+        tupleSchema: req.tupleSchema,
+        usageHint: getUsageHint?.(),
+        onResponseMetadata: (metadata) => {
+          for (const callId of metadata.functionCallIds ?? []) {
+            responseFunctionCallIds.add(callId);
+          }
+        },
+      });
+      if (result.responseId && affinityMap && conversationId) {
+        affinityMap.record(
+          result.responseId,
+          currentEntryId,
+          conversationId,
+          turnState,
+          req.codexRequest.instructions ?? undefined,
+          result.usage.input_tokens,
+          Array.from(responseFunctionCallIds),
+          variantHash,
+        );
+      }
+      if (result.usage) {
+        const u = result.usage;
+        const uncached = u.cached_tokens ? u.input_tokens - u.cached_tokens : u.input_tokens;
+        const hitPct = u.input_tokens > 0
+          ? `${((u.cached_tokens ?? 0) / u.input_tokens * 100).toFixed(1)}%`
+          : "n/a";
+        console.log(
+          `[${fmt.tag}] Account ${currentEntryId} | rid=${requestId.slice(0, 8)} | Usage: in=${u.input_tokens}` +
+          (u.cached_tokens ? ` (cached=${u.cached_tokens} uncached=${uncached})` : "") +
+          ` out=${u.output_tokens}` +
+          (u.reasoning_tokens ? ` reasoning=${u.reasoning_tokens}` : "") +
+          ` | hit=${hitPct}`,
+        );
+        if (u.input_tokens > 10_000) {
+          console.warn(`[${fmt.tag}] ⚠ High input token count: ${u.input_tokens} tokens`);
+        }
+      }
+      releaseAccount(accountPool, currentEntryId, annotateImageGenOutcome(result.usage, req.expectsImageGen), released);
+      return c.json(result.response);
+    } catch (collectErr) {
+      // Upstream FIN'd mid-reasoning (typically gpt-5.5 xhigh > 120 s cap).
+      // Cross-account retry would re-hit the same cap and burn the pool, so
+      // we fail fast with 504. The proxy can't recover this — the client
+      // needs to lower reasoning effort or pick a different model.
+      if (collectErr instanceof UpstreamPrematureCloseError) {
+        const email = accountPool.getEntry(currentEntryId)?.email ?? "?";
+        console.warn(
+          `[${fmt.tag}] Account ${currentEntryId} (${email}) | upstream premature close (hadReasoning=${collectErr.hadReasoning} events=${collectErr.eventCount}) — failing fast, not retrying`,
+        );
+        recordStreamCloseEvent({
+          kind: "upstream-premature",
+          requestId,
+          tag: fmt.tag,
+          model: req.model,
+          accountEntryId: currentEntryId,
+          variantHash,
+          responseId: collectErr.responseId,
+          eventCount: collectErr.eventCount,
+          hadReasoning: collectErr.hadReasoning,
+          detail: collectErr.message,
+        });
+        releaseAccount(accountPool, currentEntryId, annotateImageGenOutcome(undefined, req.expectsImageGen), released);
+        c.status(504);
+        return c.json(fmt.formatError(504, collectErr.message));
+      }
+
+      if (collectErr instanceof EmptyResponseError && attempt <= MAX_EMPTY_RETRIES) {
+        const email = accountPool.getEntry(currentEntryId)?.email ?? "?";
+        console.warn(
+          `[${fmt.tag}] Account ${currentEntryId} (${email}) | Empty response (attempt ${attempt}/${MAX_EMPTY_RETRIES + 1}), switching account...`,
+        );
+        accountPool.recordEmptyResponse(currentEntryId);
+        releaseAccount(accountPool, currentEntryId, annotateImageGenOutcome(collectErr.usage, req.expectsImageGen), released);
+        restoreImplicitResumeRequest?.();
+
+        const newAcquired = acquireAccount(accountPool, req.codexRequest.model, undefined, fmt.tag);
+        if (!newAcquired) {
+          c.status(502);
+          return c.json(fmt.formatError(502, "Codex returned an empty response and no other accounts are available for retry"));
+        }
+
+        currentEntryId = newAcquired.entryId;
+        currentApi = buildCodexApi(newAcquired.token, newAcquired.accountId, cookieJar, newAcquired.entryId, proxyPool);
+        setActiveAccount?.(currentEntryId, currentApi);
+        const retryStartMs = Date.now();
+        try {
+          currentRawResponse = await withRetry(
+            () => currentApi.createResponse(req.codexRequest, abortController.signal, undefined, buildPoolCtx?.(currentEntryId)),
+            { tag: fmt.tag },
+          );
+          enqueueLogEntry({
+            requestId,
+            direction: "egress",
+            method: "POST",
+            path: "/codex/responses",
+            model: req.model,
+            provider: "codex",
+            status: currentRawResponse.status,
+            latencyMs: Date.now() - retryStartMs,
+            stream: req.isStreaming,
+            request: {
+              model: req.codexRequest.model,
+              stream: req.codexRequest.stream,
+              useWebSocket: req.codexRequest.useWebSocket,
+            },
+          });
+        } catch (retryErr) {
+          releaseAccount(accountPool, currentEntryId, annotateImageGenOutcome(undefined, req.expectsImageGen), released);
+          const msg = retryErr instanceof Error ? retryErr.message : "Upstream request failed";
+          enqueueLogEntry({
+            requestId,
+            direction: "egress",
+            method: "POST",
+            path: "/codex/responses",
+            model: req.model,
+            provider: "codex",
+            status: retryErr instanceof CodexApiError ? retryErr.status : null,
+            latencyMs: Date.now() - retryStartMs,
+            stream: req.isStreaming,
+            error: msg,
+            request: {
+              model: req.codexRequest.model,
+              stream: req.codexRequest.stream,
+              useWebSocket: req.codexRequest.useWebSocket,
+            },
+          });
+          if (retryErr instanceof CodexApiError) {
+            const code = toErrorStatus(retryErr.status);
+            c.status(code);
+            return c.json(fmt.formatError(code, retryErr.message));
+          }
+          throw retryErr;
+        }
+        continue;
+      }
+
+      // Mid-SSE upstream errors (e.g. "No tool output found for function call",
+      // "previous_response_not_found") need the same strip+retry recovery as
+      // HTTP-time errors. Rethrow so the outer handleProxyRequest catch runs
+      // its unified classification once. Critically, do NOT release the slot
+      // here — outer catch's strip+retry continues on the same entryId and
+      // would race another acquirer if we released early. Outer catch is
+      // responsible for the release on the final respond/retry decision (the
+      // released Set guards against double-release on terminal paths).
+      if (collectErr instanceof CodexApiError) {
+        console.warn(
+          `[${fmt.tag}] Account ${currentEntryId} | upstream ${collectErr.status} during collect: ${stripCodexErrorPrefix(collectErr.message).slice(0, 200)}`,
+        );
+        throw collectErr;
+      }
+      releaseAccount(accountPool, currentEntryId, annotateImageGenOutcome(undefined, req.expectsImageGen), released);
+      if (collectErr instanceof EmptyResponseError) {
+        const email = accountPool.getEntry(currentEntryId)?.email ?? "?";
+        console.warn(
+          `[${fmt.tag}] Account ${currentEntryId} (${email}) | Empty response (attempt ${attempt}/${MAX_EMPTY_RETRIES + 1}), all retries exhausted`,
+        );
+        accountPool.recordEmptyResponse(currentEntryId);
+        c.status(502);
+        return c.json(fmt.formatError(502, "Codex returned empty responses across all available accounts"));
+      }
+      const msg = collectErr instanceof Error ? collectErr.message : "Unknown error";
+      const statusMatch = msg.match(/HTTP\/[\d.]+ (\d{3})/);
+      const upstreamStatus = statusMatch ? parseInt(statusMatch[1], 10) : 0;
+      const code = toErrorStatus(upstreamStatus);
+      c.status(code);
+      return c.json(fmt.formatError(code, msg));
+    }
+  }
+}
