+++ src/routes/shared/proxy-handler.ts
@@ -13,13 +13,11 @@ import type { Context } from "hono";
 import type { StatusCode } from "hono/utils/http-status";
 import { stream } from "hono/streaming";
 import {
-  CodexApi,
   CodexApiError,
   PreviousResponseWebSocketError,
 } from "../../proxy/codex-api.js";
 import type { CodexResponsesRequest } from "../../proxy/codex-api.js";
 import type { UpstreamAdapter } from "../../proxy/upstream-adapter.js";
-import { EmptyResponseError, UpstreamPrematureCloseError } from "../../translation/codex-event-extractor.js";
 import type { AccountPool } from "../../auth/account-pool.js";
 import type { CookieJar } from "../../proxy/cookie-jar.js";
 import type { ProxyPool } from "../../proxy/proxy-pool.js";
@@ -29,11 +27,13 @@ import { acquireAccount, releaseAccount } from "./account-acquisition.js";
 import { handleCodexApiError, toErrorStatus } from "./proxy-error-handler.js";
 import { isPreviousResponseNotFoundError, isUnansweredFunctionCallError } from "../../proxy/error-classification.js";
 import { streamResponse } from "./response-processor.js";
+import { handleNonStreaming } from "./non-streaming-handler.js";
+import { annotateImageGenOutcome, buildCodexApi, stripCodexErrorPrefix } from "./proxy-handler-utils.js";
 import type { UsageInfo } from "../../translation/codex-event-extractor.js";
 import { parseRateLimitHeaders, rateLimitToQuota, type ParsedRateLimit } from "../../proxy/rate-limit-headers.js";
 import { getConfig } from "../../config.js";
 import { jitterInt } from "../../utils/jitter.js";
-import { getSessionAffinityMap, type SessionAffinityMap } from "../../auth/session-affinity.js";
+import { getSessionAffinityMap } from "../../auth/session-affinity.js";
 import { enqueueLogEntry } from "../../logs/entry.js";
 import { recordStreamCloseEvent, type StreamCloseContextBase } from "../../logs/stream-close-event.js";
 import { randomUUID } from "crypto";
@@ -126,8 +126,6 @@ export interface HandleDirectRequestOptions {
   fmt: FormatAdapter;
 }
 
-const MAX_EMPTY_RETRIES = 2;
-
 /** Upper bound on how stale an implicit-resume `previous_response_id` may be.
  *  Must stay in sync with `DEFAULT_POOL_CONFIG.maxAgeMs` (3_300_000 ms) in
  *  `src/proxy/ws-pool.ts`: once the pool rotates the underlying connection,
@@ -189,32 +187,6 @@ function buildVariantIdentity(
     parts.push(`anchor:${identity.derivedConversationId}`);
   }
   return parts.length > 0 ? parts.join("\x00") : null;
-}
-
-/** Strip CodexApiError's "Codex API error (NNN): " prefix so log warns that
- *  already include status= don't duplicate it inside the message body. */
-function stripCodexErrorPrefix(msg: string): string {
-  return msg.replace(/^Codex API error \(\d+\): /, "");
-}
-
-/** Annotate a usage payload with image_generation attempt outcome before
- *  releasing the account, so `recordUsage` can split it into success vs failed
- *  counters. Synthesizes a usage object when the failure path has none. */
-function annotateImageGenOutcome(
-  usage: UsageInfo | undefined,
-  expectsImageGen: boolean | undefined,
-): UsageInfo | undefined {
-  if (!expectsImageGen) return usage;
-  const succeeded = (usage?.image_output_tokens ?? 0) > 0;
-  if (usage) {
-    return { ...usage, image_request_attempted: true, image_request_succeeded: succeeded };
-  }
-  return {
-    input_tokens: 0,
-    output_tokens: 0,
-    image_request_attempted: true,
-    image_request_succeeded: false,
-  };
 }
 
 export interface ImplicitResumeOpts {
@@ -307,17 +279,6 @@ export async function staggerIfNeeded(prevSlotMs: number | null): Promise<void> 
   const target = jitterInt(intervalMs, 0.3);
   const wait = target - elapsed;
   if (wait > 0) await new Promise((r) => setTimeout(r, wait));
-}
-
-function buildCodexApi(
-  token: string,
-  accountId: string | null,
-  cookieJar: CookieJar | undefined,
-  entryId: string,
-  proxyPool?: ProxyPool,
-): CodexApi {
-  const proxyUrl = proxyPool?.resolveProxyUrl(entryId);
-  return new CodexApi(token, accountId, cookieJar, entryId, proxyUrl);
 }
 
 function canReturnStreamError(req: ProxyRequest, fmt: FormatAdapter): boolean {
@@ -870,233 +831,6 @@ export async function handleProxyRequest(options: HandleProxyRequestOptions): Pr
       console.log(`[${fmt.tag}] Fallback → account ${retry.entryId}`);
       await staggerIfNeeded(retry.prevSlotMs);
       continue;
-    }
-  }
-}
-
-interface HandleNonStreamingOptions {
-  c: Context;
-  accountPool: AccountPool;
-  cookieJar?: CookieJar;
-  req: ProxyRequest;
-  fmt: FormatAdapter;
-  proxyPool?: ProxyPool;
-  initialApi: CodexApi;
-  initialResponse: Response;
-  initialEntryId: string;
-  abortController: AbortController;
-  released: Set<string>;
-  requestId: string;
-  affinityMap?: SessionAffinityMap;
-  conversationId?: string | null;
-  turnState?: string;
-  getUsageHint?: () => UsageHint | undefined;
-  restoreImplicitResumeRequest?: () => void;
-  buildPoolCtx?: (forEntryId: string) => WsPoolContext | undefined;
-  setActiveAccount?: (entryId: string, api: CodexApi) => void;
-  variantHash?: string;
-}
-
-async function handleNonStreaming(options: HandleNonStreamingOptions): Promise<Response> {
-  const {
-    c,
-    accountPool,
-    cookieJar,
-    req,
-    fmt,
-    proxyPool,
-    initialApi,
-    initialResponse,
-    initialEntryId,
-    abortController,
-    released,
-    requestId,
-    affinityMap,
-    conversationId,
-    turnState,
-    getUsageHint,
-    restoreImplicitResumeRequest,
-    buildPoolCtx,
-    setActiveAccount,
-    variantHash,
-  } = options;
-  let currentEntryId = initialEntryId;
-  let currentApi = initialApi;
-  let currentRawResponse = initialResponse;
-
-  for (let attempt = 1; ; attempt++) {
-    try {
-      const responseFunctionCallIds = new Set<string>();
-      const result = await fmt.collectTranslator({
-        api: currentApi,
-        response: currentRawResponse,
-        model: req.model,
-        tupleSchema: req.tupleSchema,
-        usageHint: getUsageHint?.(),
-        onResponseMetadata: (metadata) => {
-          for (const callId of metadata.functionCallIds ?? []) {
-            responseFunctionCallIds.add(callId);
-          }
-        },
-      });
-      if (result.responseId && affinityMap && conversationId) {
-        affinityMap.record(
-          result.responseId,
-          currentEntryId,
-          conversationId,
-          turnState,
-          req.codexRequest.instructions ?? undefined,
-          result.usage.input_tokens,
-          Array.from(responseFunctionCallIds),
-          variantHash,
-        );
-      }
-      if (result.usage) {
-        const u = result.usage;
-        const uncached = u.cached_tokens ? u.input_tokens - u.cached_tokens : u.input_tokens;
-        const hitPct = u.input_tokens > 0
-          ? `${((u.cached_tokens ?? 0) / u.input_tokens * 100).toFixed(1)}%`
-          : "n/a";
-        console.log(
-          `[${fmt.tag}] Account ${currentEntryId} | rid=${requestId.slice(0, 8)} | Usage: in=${u.input_tokens}` +
-          (u.cached_tokens ? ` (cached=${u.cached_tokens} uncached=${uncached})` : "") +
-          ` out=${u.output_tokens}` +
-          (u.reasoning_tokens ? ` reasoning=${u.reasoning_tokens}` : "") +
-          ` | hit=${hitPct}`,
-        );
-        if (u.input_tokens > 10_000) {
-          console.warn(`[${fmt.tag}] ⚠ High input token count: ${u.input_tokens} tokens`);
-        }
-      }
-      releaseAccount(accountPool, currentEntryId, annotateImageGenOutcome(result.usage, req.expectsImageGen), released);
-      return c.json(result.response);
-    } catch (collectErr) {
-      // Upstream FIN'd mid-reasoning (typically gpt-5.5 xhigh > 120 s cap).
-      // Cross-account retry would re-hit the same cap and burn the pool, so
-      // we fail fast with 504. The proxy can't recover this — the client
-      // needs to lower reasoning effort or pick a different model.
-      if (collectErr instanceof UpstreamPrematureCloseError) {
-        const email = accountPool.getEntry(currentEntryId)?.email ?? "?";
-        console.warn(
-          `[${fmt.tag}] Account ${currentEntryId} (${email}) | upstream premature close (hadReasoning=${collectErr.hadReasoning} events=${collectErr.eventCount}) — failing fast, not retrying`,
-        );
-        recordStreamCloseEvent({
-          kind: "upstream-premature",
-          requestId,
-          tag: fmt.tag,
-          model: req.model,
-          accountEntryId: currentEntryId,
-          variantHash,
-          responseId: collectErr.responseId,
-          eventCount: collectErr.eventCount,
-          hadReasoning: collectErr.hadReasoning,
-          detail: collectErr.message,
-        });
-        releaseAccount(accountPool, currentEntryId, annotateImageGenOutcome(undefined, req.expectsImageGen), released);
-        c.status(504);
-        return c.json(fmt.formatError(504, collectErr.message));
-      }
-
-      if (collectErr instanceof EmptyResponseError && attempt <= MAX_EMPTY_RETRIES) {
-        const email = accountPool.getEntry(currentEntryId)?.email ?? "?";
-        console.warn(
-          `[${fmt.tag}] Account ${currentEntryId} (${email}) | Empty response (attempt ${attempt}/${MAX_EMPTY_RETRIES + 1}), switching account...`,
-        );
-        accountPool.recordEmptyResponse(currentEntryId);
-        releaseAccount(accountPool, currentEntryId, annotateImageGenOutcome(collectErr.usage, req.expectsImageGen), released);
-        restoreImplicitResumeRequest?.();
-
-        const newAcquired = acquireAccount(accountPool, req.codexRequest.model, undefined, fmt.tag);
-        if (!newAcquired) {
-          c.status(502);
-          return c.json(fmt.formatError(502, "Codex returned an empty response and no other accounts are available for retry"));
-        }
-
-        currentEntryId = newAcquired.entryId;
-        currentApi = buildCodexApi(newAcquired.token, newAcquired.accountId, cookieJar, newAcquired.entryId, proxyPool);
-        setActiveAccount?.(currentEntryId, currentApi);
-        const retryStartMs = Date.now();
-        try {
-          currentRawResponse = await withRetry(
-            () => currentApi.createResponse(req.codexRequest, abortController.signal, undefined, buildPoolCtx?.(currentEntryId)),
-            { tag: fmt.tag },
-          );
-          enqueueLogEntry({
-            requestId,
-            direction: "egress",
-            method: "POST",
-            path: "/codex/responses",
-            model: req.model,
-            provider: "codex",
-            status: currentRawResponse.status,
-            latencyMs: Date.now() - retryStartMs,
-            stream: req.isStreaming,
-            request: {
-              model: req.codexRequest.model,
-              stream: req.codexRequest.stream,
-              useWebSocket: req.codexRequest.useWebSocket,
-            },
-          });
-        } catch (retryErr) {
-          releaseAccount(accountPool, currentEntryId, annotateImageGenOutcome(undefined, req.expectsImageGen), released);
-          const msg = retryErr instanceof Error ? retryErr.message : "Upstream request failed";
-          enqueueLogEntry({
-            requestId,
-            direction: "egress",
-            method: "POST",
-            path: "/codex/responses",
-            model: req.model,
-            provider: "codex",
-            status: retryErr instanceof CodexApiError ? retryErr.status : null,
-            latencyMs: Date.now() - retryStartMs,
-            stream: req.isStreaming,
-            error: msg,
-            request: {
-              model: req.codexRequest.model,
-              stream: req.codexRequest.stream,
-              useWebSocket: req.codexRequest.useWebSocket,
-            },
-          });
-          if (retryErr instanceof CodexApiError) {
-            const code = toErrorStatus(retryErr.status);
-            c.status(code);
-            return c.json(fmt.formatError(code, retryErr.message));
-          }
-          throw retryErr;
-        }
-        continue;
-      }
-
-      // Mid-SSE upstream errors (e.g. "No tool output found for function call",
-      // "previous_response_not_found") need the same strip+retry recovery as
-      // HTTP-time errors. Rethrow so the outer handleProxyRequest catch runs
-      // its unified classification once. Critically, do NOT release the slot
-      // here — outer catch's strip+retry continues on the same entryId and
-      // would race another acquirer if we released early. Outer catch is
-      // responsible for the release on the final respond/retry decision (the
-      // released Set guards against double-release on terminal paths).
-      if (collectErr instanceof CodexApiError) {
-        console.warn(
-          `[${fmt.tag}] Account ${currentEntryId} | upstream ${collectErr.status} during collect: ${stripCodexErrorPrefix(collectErr.message).slice(0, 200)}`,
-        );
-        throw collectErr;
-      }
-      releaseAccount(accountPool, currentEntryId, annotateImageGenOutcome(undefined, req.expectsImageGen), released);
-      if (collectErr instanceof EmptyResponseError) {
-        const email = accountPool.getEntry(currentEntryId)?.email ?? "?";
-        console.warn(
-          `[${fmt.tag}] Account ${currentEntryId} (${email}) | Empty response (attempt ${attempt}/${MAX_EMPTY_RETRIES + 1}), all retries exhausted`,
-        );
-        accountPool.recordEmptyResponse(currentEntryId);
-        c.status(502);
-        return c.json(fmt.formatError(502, "Codex returned empty responses across all available accounts"));
-      }
-      const msg = collectErr instanceof Error ? collectErr.message : "Unknown error";
-      const statusMatch = msg.match(/HTTP\/[\d.]+ (\d{3})/);
-      const upstreamStatus = statusMatch ? parseInt(statusMatch[1], 10) : 0;
-      const code = toErrorStatus(upstreamStatus);
-      c.status(code);
-      return c.json(fmt.formatError(code, msg));
     }
   }
 }
